Fix for BadRequestError to make them visible to client
When attempting to make a PATCH request to /api/v1/study/1/samples, with
obviously bad parameters, an exception was being raised server-side, and
therefore a descriptive error message was not being returned to the user
from qiita_client.

Bug appears to be the assumption that 'error_description' is present as
a key in r_json. A check has been added for the key.